Version uploaded to DDS 3 and DDS 4 on 27.09.2019
This is the version that is running on the current DDSs. The ramps look like they are working, although there is some timing issue at the end of the hardware (HW) ramps, which requires us to give it a bit more time, it cannot immediately go to the next step. The frequency jumps to the starting frequency in this manner.
NOTE!!! The interrupt pin is on Digital 2 on DDS 3 and DDS 4 but it's on Digital 9 on the new DDS